# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Almuzini Children Hospital EHR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/summary:
    get:
      operationId: getPatientSummary
      tags: [patients]
      summary: Get patient clinical summary
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Patient summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientSummary"

  # APPOINTMENTS
...
  /lab-orders:
    get:
      operationId: listLabOrders
      tags: [lab]
      summary: List lab orders
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lab orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabOrder"
    post:
      operationId: createLabOrder
      tags: [lab]
      summary: Create lab order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab-orders/{id}:
    get:
      operationId: getLabOrder
      tags: [lab]
      summary: Get lab order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lab order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
    patch:
      operationId: updateLabOrder
      tags: [lab]
      summary: Update lab order / add results
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"

  # RADIOLOGY
  /radiology-orders:
    get:
      operationId: listRadiologyOrders
      tags: [radiology]
      summary: List radiology orders
      parameters:
        - name: patientId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Radiology orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RadiologyOrder"
    post:
      operationId: createRadiologyOrder
      tags: [radiology]
      summary: Create radiology order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiologyOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiologyOrder"
  /radiology-orders/{id}:
    patch:
      operationId: updateRadiologyOrder
      tags: [radiology]
      summary: Update radiology order / add report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiologyOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiologyOrder"

  # BILLING
...
  /drugs:
    get:
      operationId: listDrugs
      tags: [pharmacy]
      summary: List drugs in inventory
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: lowStock
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Drugs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drug"
    post:
      operationId: createDrug
      tags: [pharmacy]
      summary: Add drug to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrugInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drug"
  /drugs/{id}:
    patch:
      operationId: updateDrug
      tags: [pharmacy]
      summary: Update drug
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrugUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drug"

  # STAFF / USERS
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, mrn, nameEn, dateOfBirth, gender, createdAt]
      properties:
        id: { type: integer }
        mrn: { type: string }
        nameEn: { type: string }
        nameAr: { type: ["string", "null"] }
        dateOfBirth: { type: string }
        gender: { type: string }
        bloodGroup: { type: ["string", "null"] }
        nationality: { type: ["string", "null"] }
        nationalId: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        guardianName: { type: ["string", "null"] }
        guardianRelation: { type: ["string", "null"] }
        guardianPhone: { type: ["string", "null"] }
        allergies: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    PatientInput:
      type: object
      required: [nameEn, dateOfBirth, gender]
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        dateOfBirth: { type: string }
        gender: { type: string }
        bloodGroup: { type: string }
        nationality: { type: string }
        nationalId: { type: string }
        phone: { type: string }
        address: { type: string }
        guardianName: { type: string }
        guardianRelation: { type: string }
        guardianPhone: { type: string }
        allergies: { type: string }
    PatientUpdate:
      type: object
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        phone: { type: string }
        address: { type: string }
        guardianName: { type: string }
        guardianRelation: { type: string }
        guardianPhone: { type: string }
        allergies: { type: string }
        bloodGroup: { type: string }
        status: { type: string }
    PatientListResponse:
      type: object
      required: [patients, total, page, limit]
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    PatientSummary:
      type: object
      required: [patient]
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        recentNotes:
          type: array
          items:
            $ref: "#/components/schemas/ClinicalNote"
        activePrescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
        pendingLabOrders:
          type: array
          items:
            $ref: "#/components/schemas/LabOrder"
        upcomingAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        latestGrowth:
          $ref: "#/components/schemas/GrowthRecord"
        vaccinations:
          type: array
          items:
            $ref: "#/components/schemas/Vaccination"

    # APPOINTMENTS
    Appointment:
      type: object
      required: [id, patientId, doctorId, scheduledAt, type, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: ["string", "null"] }
        doctorId: { type: integer }
        doctorName: { type: ["string", "null"] }
        scheduledAt: { type: string }
        duration: { type: integer }
        type: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        chiefComplaint: { type: ["string", "null"] }
        createdAt: { type: string }
...
    ClinicalNote:
      type: object
      required: [id, patientId, authorId, type, content, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        authorId: { type: integer }
        authorName: { type: ["string", "null"] }
        type: { type: string }
        subjective: { type: ["string", "null"] }
        objective: { type: ["string", "null"] }
        assessment: { type: ["string", "null"] }
        plan: { type: ["string", "null"] }
        content: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Prescription:
      type: object
      required: [id, patientId, prescriberId, drugName, dosage, frequency, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: ["string", "null"] }
        prescriberId: { type: integer }
        prescriberName: { type: ["string", "null"] }
        drugName: { type: string }
        drugNameAr: { type: ["string", "null"] }
        dosage: { type: string }
        frequency: { type: string }
        duration: { type: ["string", "null"] }
        route: { type: ["string", "null"] }
        instructions: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
...
    LabOrder:
      type: object
      required: [id, patientId, orderedById, testName, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: ["string", "null"] }
        orderedById: { type: integer }
        orderedByName: { type: ["string", "null"] }
        testName: { type: string }
        testCode: { type: ["string", "null"] }
        priority: { type: string }
        status: { type: string }
        result: { type: ["string", "null"] }
        resultValue: { type: ["string", "null"] }
        unit: { type: ["string", "null"] }
        referenceRange: { type: ["string", "null"] }
        isCritical: { type: boolean }
        notes: { type: ["string", "null"] }
        collectedAt: { type: ["string", "null"] }
        resultedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    LabOrderInput:
      type: object
      required: [patientId, testName]
      properties:
        patientId: { type: integer }
        testName: { type: string }
        testCode: { type: string }
        priority: { type: string }
        notes: { type: string }
    LabOrderUpdate:
      type: object
      properties:
        status: { type: string }
        result: { type: string }
        resultValue: { type: string }
        unit: { type: string }
        referenceRange: { type: string }
        isCritical: { type: boolean }
        collectedAt: { type: string }
        resultedAt: { type: string }

    # RADIOLOGY
    RadiologyOrder:
      type: object
      required: [id, patientId, orderedById, modality, studyDescription, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: ["string", "null"] }
        orderedById: { type: integer }
        orderedByName: { type: ["string", "null"] }
        modality: { type: string }
        studyDescription: { type: string }
        priority: { type: string }
        status: { type: string }
        report: { type: ["string", "null"] }
        radiologistId: { type: ["integer", "null"] }
        scheduledAt: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    RadiologyOrderInput:
      type: object
      required: [patientId, modality, studyDescription]
      properties:
        patientId: { type: integer }
        modality: { type: string }
        studyDescription: { type: string }
        priority: { type: string }
        scheduledAt: { type: string }
    RadiologyOrderUpdate:
      type: object
      properties:
        status: { type: string }
        report: { type: string }
        radiologistId: { type: integer }
        scheduledAt: { type: string }
        completedAt: { type: string }

    # BILLING
...
    Drug:
      type: object
      required: [id, name, category, stockQuantity, unit, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        genericName: { type: ["string", "null"] }
        category: { type: string }
        stockQuantity: { type: integer }
        minStockLevel: { type: integer }
        unit: { type: string }
        unitPrice: { type: number }
        batchNumber: { type: ["string", "null"] }
        expiryDate: { type: ["string", "null"] }
        manufacturer: { type: ["string", "null"] }
        isControlled: { type: boolean }
        createdAt: { type: string }
    DrugInput:
      type: object
      required: [name, category, stockQuantity, unit]
      properties:
        name: { type: string }
        nameAr: { type: string }
        genericName: { type: string }
        category: { type: string }
        stockQuantity: { type: integer }
        minStockLevel: { type: integer }
        unit: { type: string }
        unitPrice: { type: number }
        batchNumber: { type: string }
        expiryDate: { type: string }
        manufacturer: { type: string }
        isControlled: { type: boolean }
    DrugUpdate:
      type: object
      properties:
        stockQuantity: { type: integer }
        unitPrice: { type: number }
        expiryDate: { type: string }
        isControlled: { type: boolean }

    # STAFF
...
    Vaccination:
      type: object
      required: [id, patientId, vaccineName, administeredDate, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        vaccineName: { type: string }
        vaccineNameAr: { type: ["string", "null"] }
        doseNumber: { type: ["integer", "null"] }
        administeredDate: { type: string }
        nextDueDate: { type: ["string", "null"] }
        batchNumber: { type: ["string", "null"] }
        administeredById: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
...
    GrowthRecord:
      type: object
      required: [id, patientId, measurementDate, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        measurementDate: { type: string }
        weight: { type: ["number", "null"] }
        height: { type: ["number", "null"] }
        headCircumference: { type: ["number", "null"] }
        bmi: { type: ["number", "null"] }
        weightPercentile: { type: ["number", "null"] }
        heightPercentile: { type: ["number", "null"] }
        bmiPercentile: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
...