# lib/api-zod/src/generated/types/labOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Almuzini Children Hospital EHR API
 * OpenAPI spec version: 0.1.0
 */

export interface LabOrder {
  id: number;
  patientId: number;
  /** @nullable */
  patientName?: string | null;
  orderedById: number;
  /** @nullable */
  orderedByName?: string | null;
  testName: string;
  /** @nullable */
  testCode?: string | null;
  priority?: string;
  status: string;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  resultValue?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  referenceRange?: string | null;
  isCritical?: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  collectedAt?: string | null;
  /** @nullable */
  resultedAt?: string | null;
  createdAt: string;
}