# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Almuzini Children Hospital EHR API
 * OpenAPI spec version: 0.1.0
 */

export interface Prescription {
  id: number;
  patientId: number;
  /** @nullable */
  patientName?: string | null;
  prescriberId: number;
  /** @nullable */
  prescriberName?: string | null;
  drugName: string;
  /** @nullable */
  drugNameAr?: string | null;
  dosage: string;
  frequency: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  instructions?: string | null;
  status: string;
  createdAt: string;
}