# lib/api-zod/src/generated/types/radiologyOrderUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Almuzini Children Hospital EHR API
 * OpenAPI spec version: 0.1.0
 */

export interface RadiologyOrderUpdate {
  status?: string;
  report?: string;
  radiologistId?: number;
  scheduledAt?: string;
  completedAt?: string;
}